+++ src/web/client/common/interfaces.ts
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License. See License.txt in the project root for license information.
+ */
+
+export interface IEntityRequestUrl {
+    requestUrl: string;
+    entityName: string;
+}
+
+export interface IAttributePath {
+    source: string;
+    relativePath: string;
+}
